backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
pinecone-client==3.0.0
openai==1.3.7
httpx==0.25.2
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
rich==13.7.0